fix: enable read-write database connections for user auth
- Change DuckDB connection from read-only to read-write mode
- Fixes authentication endpoints that need to write to usuarios table
- Required for /api/auth/init-db and user creation operations

+++ backend/models/pedidos_sugeridos.py
@@ -525,9 +525,9 @@ class PedidoSugeridoResumen(BaseModel):
     # Totales
     total_productos: int
     total_lineas: int
-    total_bultos: Decimal
-    total_unidades: Decimal
-    total_peso_kg: Optional[Decimal] = None
+    total_bultos: float
+    total_unidades: float
+    total_peso_kg: Optional[float] = None
 
     # Fechas importantes
     fecha_entrega_solicitada: Optional[date] = None
